auth-requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.32.0
pydantic-settings==2.6.0
pyjwt[crypto]==2.9.0
redis==5.2.0
requests==2.32.3
pydantic[email]
python-multipart==0.0.12
httpx==0.27.2
Authlib==1.3.2